[appveyor.yml] use %WITH_COMPILER% to pip install cython
Previously, we were relying on the fact that the PyPI server contained
precompiled Windows wheels for Cython. However, it seems that this is no
longer the case (I'm not sure if it's just temporary). So for now we have
to compile Cython from the source *.tar.gz archives that are available to
pip. These need to be built using the correct MSVC compiler and/or Windows
SDK, so we call pip install cython through %WITH_COMPILER% that sets the
correct environment variables.