Remove wheel from setup_requires
When downloading the source and building this locally python-wheel is not required. It is only required for developers who build bdist packages for publishing. The CI jobs that do this already manually install wheel anyway and most documentation I see assumes that will be the case. For distro packagers, having this in here just means we have to include an unused dependency or patch it out.